use assembly keccak for amd64
